Update jQuery Smooth Scroll to v1.5.5. Fix #71.
