+++ djfrontend/templatetags/settings.py
@@ -9,5 +9,5 @@ DJFRONTEND_JQUERYUI_DEFAULT = '1.11.1'
 DJFRONTEND_JQUERY_DATATABLES_VERSION_DEFAULT = '1.10.2'
 DJFRONTEND_JQUERY_FORMSET_DEFAULT = '1.2'
 DJFRONTEND_JQUERY_SCROLLTO_DEFAULT = '1.4.13'
-DJFRONTEND_JQUERY_SMOOTHSCROLL_DEFAULT = '1.5.0'
+DJFRONTEND_JQUERY_SMOOTHSCROLL_DEFAULT = '1.5.5'
 DJFRONTEND_TWBS_VERSION_DEFAULT = '3.2.0'
